requirements.txt
# Core AI / LLM
llama-index-core>=0.12.0
llama-index-llms-cerebras>=0.3.0
llama-index-embeddings-huggingface>=0.4.0

# Structured output
pydantic>=2.0.0

# Environment & config
python-dotenv>=1.0.0

# Web UI
streamlit>=1.40.0
pandas>=2.0.0

# Database
pymongo>=4.6.0
dnspython>=2.4.0
certifi>=2024.0.0

# Utilities
gitpython>=3.1.0
